feat: add journal deletion logic using archive shelf (#87)
* feat: add button for showing journal options

* feat: add dialog for deleting journals

* feat: add connection between server and client for deleting journals

* feat: add soft deletes for journals by using archives

* feat: add separate logic for rendering archive shelves and normal shelves

* chore: apply pr review comments

* fix: typo

+++ journal-service/src/main/java/com/kioke/journal/service/JournalService.java
@@ -5,6 +5,7 @@ import com.kioke.journal.model.Journal;
 import com.kioke.journal.model.Shelf;
 import com.kioke.journal.model.User;
 import com.kioke.journal.repository.JournalRepository;
+import com.kioke.journal.repository.ShelfRepository;
 import java.util.ArrayList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
@@ -13,18 +14,38 @@ import org.springframework.stereotype.Service;
 @Service
 public class JournalService {
   @Autowired @Lazy JournalRepository journalRepository;
+  @Autowired @Lazy ShelfRepository shelfRepository;
 
   public Journal getJournalById(String jid) throws JournalNotFoundException {
-    return journalRepository.findById(jid).orElseThrow(() -> new JournalNotFoundException(jid));
+    return journalRepository.findById(jid).orElseThrow(() -> new JournalNotFoundException());
   }
 
   public Journal createJournal(User user, Shelf shelf, String title) {
-    Journal journal = Journal.builder().title(title).users(new ArrayList<>()).shelf(shelf).build();
+    Journal journal =
+        Journal.builder()
+            .title(title)
+            .users(new ArrayList<>())
+            .shelf(shelf)
+            .isDeleted(false)
+            .build();
+
     journal = journalRepository.save(journal);
     return journal;
   }
 
-  public void deleteJournal(Journal journal) {
-    journalRepository.delete(journal);
+  public void deleteJournal(User user, Journal journal) {
+    if (journal.isDeleted()) {
+      journalRepository.delete(journal);
+      return;
+    }
+
+    Shelf archive =
+        shelfRepository
+            .findByOwnerAndIsArchiveTrue(user)
+            .orElseThrow(() -> new IllegalStateException("User does not have archive shelf."));
+
+    journal.setDeleted(true);
+    journal.setShelf(archive);
+    journalRepository.save(journal);
   }
 }
